JINRO_PROJ/backend/requirements.txt
fastapi==0.110.3
uvicorn[standard]==0.30.1
python-multipart==0.0.9
itsdangerous==2.2.0

pydantic==2.7.1
email-validator==2.2.0
typing_extensions>=4.10.0

sqlalchemy==2.0.36
pymysql==1.1.1
httpx==0.28.1
requests==2.32.3
python-dotenv==1.0.1
